refactor MultiLineCellRenderer to separate file
