+++ java/src/jmri/util/swing/MultiLineCellRenderer.java
@@ -1,0 +1,41 @@
+package jmri.util.swing;
+
+import java.awt.Component;
+
+import javax.swing.JTable;
+import javax.swing.JTextArea;
+import javax.swing.table.TableCellRenderer;
+import javax.swing.UIManager;
+
+/**
+ * Renderer to display multiple lines in a JTable cell
+ */
+public class MultiLineCellRenderer extends JTextArea implements TableCellRenderer {
+
+    public MultiLineCellRenderer() {
+        setLineWrap(true);
+        setWrapStyleWord(true);
+        setOpaque(true);
+    }
+
+    @Override
+    public Component getTableCellRendererComponent(JTable table, Object value,
+        boolean isSelected, boolean hasFocus, int row, int column) {
+        if (isSelected) {
+            setForeground(table.getSelectionForeground());
+            setBackground(table.getSelectionBackground());
+        } else {
+            setForeground(table.getForeground());
+            setBackground(table.getBackground());
+        }
+        setFont(table.getFont());
+        if (hasFocus) {
+            if (table.isCellEditable(row, column)) {
+                setForeground(UIManager.getColor("Table.focusCellForeground"));
+                setBackground(UIManager.getColor("Table.focusCellBackground"));
+            }
+        }
+        setText((value == null) ? "" : value.toString());
+        return this;
+    }
+}
